Fix submit-form 422: scale answers (rating/nps/csat/opinionScale) must be integers
Same class as the completionTimeInSeconds fix, now audited against the
backend's actual Rust schema (core-api/src/models/submit.rs) instead of
field-by-field: every numeric field in the submit-form tree is an integer
type (i32/i64) except fileUpload[].fileSizeMb (f64). The SDKs typed the
four scale answers and videoAudio.durationSeconds as Double — kotlinx
emits whole doubles as "4.0" -> 422 on Android/KMP/Compose.

Now Int across core/kmp-sdk/ios-native (Answer types + toNum builder
helpers, which round). Already-correct per the audit: timeSpentMs (i64 ->
Long), fileSizeMb (f64 -> Double), number/amount as strings, ratingMatrix
as raw JSON passthrough. dist/ rebuilt.

+++ core/src/commonMain/kotlin/com/encatch/core/Answer.kt
@@ -66,7 +66,7 @@ data class VideoAudioAnswer(
     val mode: String,
     val fileUrl: String? = null,
     val text: String? = null,
-    val durationSeconds: Double? = null,
+    val durationSeconds: Int? = null,
     val transcriptText: String? = null,
 )
 
@@ -119,12 +119,14 @@ data class PaymentsUpiAnswer(
  */
 @Serializable
 data class Answer(
-    val nps: Double? = null,
+    // Int, not Double: the backend deserializes scale answers as i32 and 422s on "4.0"
+    // (see completionTimeInSeconds' twin comment).
+    val nps: Int? = null,
     val nestedSelection: List<String>? = null,
     val longText: String? = null,
     val shortAnswer: String? = null,
     val singleChoice: String? = null,
-    val rating: Double? = null,
+    val rating: Int? = null,
     val yesNo: Boolean? = null,
     val consent: Boolean? = null,
     val multipleChoiceMultiple: List<String>? = null,
@@ -136,8 +138,8 @@ data class Answer(
     val matrixMultipleChoice: Map<String, List<String>>? = null,
     val others: String? = null,
     val date: String? = null,
-    val csat: Double? = null,
-    val opinionScale: Double? = null,
+    val csat: Int? = null,
+    val opinionScale: Int? = null,
     val ranking: List<String>? = null,
     val pictureChoice: List<String>? = null,
     val pictureChoiceOther: String? = null,
